move actual wizards to wizards.py
